The code commit includes changes in the following files:
1. dist/index.js
2. dist/index.js.map
3. index.js
4. response.js

The changes in the index.js file involve replacing `core.info` with `core.debug` for logging the prompts and input values. This change is made to use the debug level of logging instead of the info level.

In the response.js file, two `console.info` statements are commented out to disable logging the request and output information.